+++ packages/workflow/src/agent/function-agent.ts
@@ -1,4 +1,4 @@
-import type { WorkflowContext } from "@llama-flow/core";
+import { type WorkflowContext, type WorkflowEvent } from "@llama-flow/core";
 import type { JSONObject } from "@llamaindex/core/global";
 import { Settings } from "@llamaindex/core/global";
 import {
@@ -7,6 +7,9 @@ import {
   type ChatMessage,
   type ChatResponseChunk,
 } from "@llamaindex/core/llms";
+import { tool } from "@llamaindex/core/tools";
+import { z } from "zod";
+import { zodToJsonSchema } from "zod-to-json-schema";
 import { AgentWorkflow } from "./agent-workflow";
 import { type AgentWorkflowState, type BaseWorkflowAgent } from "./base";
 import {
@@ -18,6 +21,51 @@ import {
 
 const DEFAULT_SYSTEM_PROMPT =
   "You are a helpful assistant. Use the provided tools to answer questions.";
+
+const STEP_HANDLER_SYSTEM_PROMPT_TPL = `
+You are a part of a workflow program that is composed of multiple steps. 
+Your task is to handle the step using the provided tools and finally send an output event back to the workflow and summarize the result.
+
+## Instructions
+### Follow these default instructions:
+1. Provide a plan to handle the actions based on context and the user request.
+2. Use the provided tools to proceed with your actions. You can call multiple tools to handle the step and send events.
+3. Always return at least one result event at the end to the workflow by using the provided result tools. Identify the event to send based on the user's instructions.
+
+### Here is the user's instructions:
+{instructions}
+`;
+
+export type ZodEvent = WorkflowEvent<unknown> & {
+  schema: z.ZodType<unknown>;
+};
+
+export type StepHandlerParams = {
+  /**
+   * Workflow context
+   */
+  workflowContext: WorkflowContext;
+  /**
+   * User instructions to guide the agent to handle the step.
+   */
+  instructions: string;
+  /**
+   * Event that this agent will return
+   */
+  results: ZodEvent[];
+  /**
+   * List of additional events that the agent can emit
+   */
+  events?: ZodEvent[];
+  /**
+   * LLM to use for the agent, required.
+   */
+  llm?: ToolCallLLM | undefined;
+  /**
+   * List of tools that the agent can use
+   */
+  tools?: BaseToolWithCall[] | undefined;
+};
 
 export type FunctionAgentParams = {
   /**
@@ -36,7 +84,7 @@ export type FunctionAgentParams = {
   /**
    * List of tools that the agent can use, requires at least one tool.
    */
-  tools: BaseToolWithCall[];
+  tools?: BaseToolWithCall[] | undefined;
   /**
    * List of agents that this agent can delegate tasks to
    * Can be a list of agent names as strings, BaseWorkflowAgent instances, or AgentWorkflow instances
@@ -46,6 +94,11 @@ export type FunctionAgentParams = {
    * Custom system prompt for the agent
    */
   systemPrompt?: string | undefined;
+};
+
+export type EmitEvent = {
+  event: WorkflowEvent<unknown> & { schema: z.ZodType<unknown> };
+  name: string;
 };
 
 export class FunctionAgent implements BaseWorkflowAgent {
@@ -72,40 +125,43 @@ export class FunctionAgent implements BaseWorkflowAgent {
     this.description =
       description ??
       "A single agent that uses the provided tools or functions.";
-    this.tools = tools;
-    if (tools.length === 0) {
-      throw new Error("FunctionAgent must have at least one tool");
-    }
-    // Process canHandoffTo to extract agent names
-    this.canHandoffTo = [];
-    if (canHandoffTo) {
-      if (Array.isArray(canHandoffTo)) {
-        if (canHandoffTo.length > 0) {
-          if (typeof canHandoffTo[0] === "string") {
+    this.tools = tools ?? [];
+    this.systemPrompt = systemPrompt ?? DEFAULT_SYSTEM_PROMPT;
+    this.canHandoffTo = this.initHandOffNames(canHandoffTo ?? []);
+  }
+
+  private initHandOffNames(
+    handoffTo: string[] | BaseWorkflowAgent[] | AgentWorkflow[],
+  ): string[] {
+    const handoffToNames: string[] = [];
+    if (handoffTo) {
+      if (Array.isArray(handoffTo)) {
+        if (handoffTo.length > 0) {
+          if (typeof handoffTo[0] === "string") {
             // string[] case
-            this.canHandoffTo = canHandoffTo as string[];
-          } else if (canHandoffTo[0] instanceof AgentWorkflow) {
+            handoffToNames.push(...(handoffTo as string[]));
+          } else if (handoffTo[0] instanceof AgentWorkflow) {
             // AgentWorkflow[] case
-            const workflows = canHandoffTo as AgentWorkflow[];
+            const workflows = handoffTo as AgentWorkflow[];
             workflows.forEach((workflow) => {
               const agentNames = workflow
                 .getAgents()
                 .map((agent) => agent.name);
-              this.canHandoffTo.push(...agentNames);
+              handoffToNames.push(...agentNames);
             });
           } else {
             // BaseWorkflowAgent[] case
-            const agents = canHandoffTo as BaseWorkflowAgent[];
-            this.canHandoffTo = agents.map((agent) => agent.name);
+            const agents = handoffTo as BaseWorkflowAgent[];
+            handoffToNames.push(...agents.map((agent) => agent.name));
           }
         }
       }
     }
-    const uniqueHandoffAgents = new Set(this.canHandoffTo);
-    if (uniqueHandoffAgents.size !== this.canHandoffTo.length) {
+    const uniqueHandoffAgents = new Set(handoffToNames);
+    if (uniqueHandoffAgents.size !== handoffToNames.length) {
       throw new Error("Duplicate handoff agents");
     }
-    this.systemPrompt = systemPrompt ?? DEFAULT_SYSTEM_PROMPT;
+    return handoffToNames;
   }
 
   async takeStep(
@@ -256,4 +312,125 @@ export class FunctionAgent implements BaseWorkflowAgent {
 
     return toolCalls;
   }
+
+  /**
+   * Create a FunctionAgent to handle a step of the workflow.
+   * @param params.workflowContext - The workflow context.
+   * @param params.result - The event to send when the agent is done.
+   * @param params.events - Additional events that the agent can emit.
+   * @param params.instructions - The user instructions to guide the agent to handle the step.
+   * @param params.tools - The tools to use for the agent.
+   * @returns A new FunctionAgent instance
+   */
+  static fromWorkflowStep({
+    workflowContext,
+    results,
+    events,
+    instructions,
+    tools,
+    llm,
+  }: StepHandlerParams): FunctionAgent {
+    if (!workflowContext) {
+      throw new Error("workflowContext must be provided");
+    }
+    if (results.length === 0) {
+      throw new Error("results must have at least one event");
+    }
+    if (!instructions) {
+      throw new Error("instructions must be provided");
+    }
+
+    // Provided tools
+    const allTools = [...(tools ?? [])];
+    // Add tools for result events
+    results.forEach((result) => {
+      if (!result.debugLabel) {
+        throw new Error("Result event must have a debug label");
+      }
+      const description = `Use this tool to send the ${result.debugLabel} event as the final result of your task.`;
+      allTools.push(
+        createEventEmitterTool(
+          `send_${result.debugLabel}`,
+          result,
+          workflowContext,
+          description,
+        ),
+      );
+    });
+    // Add tools for additional events
+    events?.forEach((event) => {
+      if (!event.debugLabel) {
+        throw new Error("Event must have a debug label");
+      }
+      allTools.push(
+        createEventEmitterTool(
+          `send_${event.debugLabel}`,
+          event,
+          workflowContext,
+          `Use this tool to send the ${event.debugLabel} event to the workflow program.`,
+        ),
+      );
+    });
+
+    // Construct the system prompt
+    const newSystemPrompt = STEP_HANDLER_SYSTEM_PROMPT_TPL.replace(
+      "{instructions}",
+      instructions,
+    );
+
+    // Check if llm is provided or default LLM is a tool call LLM
+    const llmToUse = llm ?? (Settings.llm as ToolCallLLM);
+    if (!llmToUse.supportToolCall) {
+      throw new Error("LLM must support tool calls");
+    }
+    // Create the function agent
+    return new FunctionAgent({
+      llm: llmToUse,
+      systemPrompt: newSystemPrompt,
+      tools: allTools,
+    });
+  }
 }
+
+/**
+ * Create a tool that sends an event to the workflow.
+ * @param name - The name of the tool.
+ * @param event - The event to send.
+ * @param workflowContext - The workflow context.
+ * @param description - The description of the tool.
+ */
+const createEventEmitterTool = (
+  name: string,
+  event: WorkflowEvent<unknown> & { schema: z.ZodType<unknown> },
+  workflowContext: WorkflowContext,
+  description?: string,
+) => {
+  // To ensure the model correctly interprets the event data, including the schema in the tool description is crucial.
+  // This is particularly important for special types like literals and enums, which the model might struggle with otherwise.
+  // By incorporating the schema into the tool description, we can facilitate the model's understanding of the event data.
+  const toolDescriptionWithSchema =
+    (description ??
+      event.schema.description ??
+      "Use this tool to send the event to the workflow.") +
+    `\n\nPlease provide the event data in the following JSON schema: ${JSON.stringify(
+      zodToJsonSchema(z.object({ eventData: event.schema })),
+    )}`;
+  return tool({
+    name: name,
+    description: toolDescriptionWithSchema,
+    parameters: z.object({
+      eventData: event.schema,
+    }),
+    execute: (
+      { eventData }: { eventData?: z.infer<typeof event.schema> },
+      getContext?: () => WorkflowContext,
+    ) => {
+      if (!getContext) {
+        throw new Error("Workflow context is not provided.");
+      }
+      const context = getContext();
+      context.sendEvent(event.with(eventData ?? {}));
+      return `Successfully sent a ${name} event!`;
+    },
+  }).bind(() => workflowContext);
+};