+++ e2e/examples/nextjs-node-runtime/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @llamaindex/next-node-runtime
 
+## 0.1.36
+
+### Patch Changes
+
+- llamaindex@0.11.8
+- @llamaindex/huggingface@0.1.14
+- @llamaindex/readers@3.1.8
+
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies [3c857f4]
+  - llamaindex@0.11.7
+
 ## 0.1.34
 
 ### Patch Changes