+++ e2e/examples/vite-import-llamaindex/CHANGELOG.md
@@ -1,5 +1,18 @@
 # vite-import-llamaindex
 
+## 0.0.35
+
+### Patch Changes
+
+- llamaindex@0.11.8
+
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [3c857f4]
+  - llamaindex@0.11.7
+
 ## 0.0.33
 
 ### Patch Changes