+++ packages/autotool/examples/01_node/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @llamaindex/autotool-01-node-example
 
+## 0.0.116
+
+### Patch Changes
+
+- llamaindex@0.11.8
+- @llamaindex/autotool@8.0.8
+
+## 0.0.115
+
+### Patch Changes
+
+- Updated dependencies [3c857f4]
+  - llamaindex@0.11.7
+  - @llamaindex/autotool@8.0.7
+
 ## 0.0.114
 
 ### Patch Changes