+++ packages/providers/assemblyai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/assemblyai
 
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [1b5af14]
+  - @llamaindex/core@0.6.10
+
 ## 0.1.8
 
 ### Patch Changes