+++ packages/providers/replicate/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/replicate
 
+## 0.0.52
+
+### Patch Changes
+
+- Updated dependencies [1b5af14]
+  - @llamaindex/core@0.6.10
+
 ## 0.0.51
 
 ### Patch Changes