+++ packages/providers/storage/qdrant/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/qdrant
 
+## 0.1.20
+
+### Patch Changes
+
+- Updated dependencies [1b5af14]
+  - @llamaindex/core@0.6.10
+
 ## 0.1.19
 
 ### Patch Changes