+++ packages/providers/voyage-ai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/voyage-ai
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [1b5af14]
+  - @llamaindex/core@0.6.10
+
 ## 1.0.15
 
 ### Patch Changes