+++ packages/tools/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @llamaindex/tools
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [1b5af14]
+  - @llamaindex/core@0.6.10
+
+## 0.0.15
+
+### Patch Changes
+
+- 3c857f4: chore: move ajv to dev deps
+
 ## 0.0.14
 
 ### Patch Changes